requirements.txt
Django==4.1
python-dotenv==1.0.0
spotipy==2.23.0
